Downgraded to DotNet5

+++ utils/logger/KLog.cs
@@ -8,7 +8,7 @@ namespace dc.assignment.primenumbers.utils.logger{
         [BsonRepresentation(BsonType.ObjectId)]        
         public string? Id { get; set; }
         public string timestamp { get; set; } = null!;
-        public Int64 nodeId { get; set; }
+        public int nodeId { get; set; }
         public string nodeAddress { get; set; } = null!;
         public string message { get; set; } = null!;
     }
